refactor: common 디렉토리 제거
