+++ apps/spectator/components/ErrorBoundary/index.tsx
