+++ apps/spectator/components/Panel/Panel.css.ts
