+++ apps/spectator/components/cheertalk/Modal/CheerTalkModal.tsx
@@ -5,8 +5,8 @@ import Banner from 'components/cheertalk/Modal/Banner';
 import CheerTalkForm from 'components/cheertalk/Modal/CheerTalkForm';
 import CheerTalkList from 'components/cheertalk/Modal/CheerTalkList';
 
-import AsyncBoundary from '@/components/common/AsyncBoundary';
-import Loader from '@/components/common/Loader';
+import AsyncBoundary from '@/components/AsyncBoundary';
+import Loader from '@/components/Loader';
 import useSocket from '@/hooks/useSocket';
 import GameCheerTalkFetcher from '@/queries/useGameCheerTalkById/Fetcher';
 import useSaveCheerTalkMutation from '@/queries/useSaveCheerTalkMutation/query';
