--- apps/spectator/components/common/GameCard/pieces/Label.css.ts
@@ -1,13 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const labelWrapper = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  fontSize: '0.875rem',
-  lineHeight: '1.25rem',
-});
-
-export const labelContent = style({
-  textAlign: 'right',
-});
