--- apps/spectator/components/common/GameCard/pieces/Score.css.ts
@@ -1,6 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const score = style({
-  fontSize: '1.875rem',
-  lineHeight: '2.25rem',
-});
