--- apps/spectator/components/common/GameCard/pieces/Status.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const status = style({
-  zIndex: 10,
-  padding: '0 1.25rem',
-  color: 'white',
-});
