--- apps/spectator/components/common/GameCard/pieces/Status.tsx
@@ -1,15 +1,0 @@
-import { clsx } from 'clsx';
-
-import { useGameCardContext } from '@/hooks/useGameCardContext';
-
-import * as styles from './Status.css';
-
-type StatusProps = {
-  className?: string;
-};
-
-export default function Status({ className }: StatusProps) {
-  const { gameQuarter } = useGameCardContext();
-
-  return <div className={clsx(className, styles.status)}>{gameQuarter}</div>;
-}
