--- apps/spectator/components/common/GameCard/pieces/Team.css.ts
@@ -1,8 +1,0 @@
-import { theme } from '@hcc/styles';
-import { style } from '@vanilla-extract/css';
-
-export const teamLogoDefault = style({
-  width: '4rem',
-  height: '4rem',
-  color: theme.colors.gray[2],
-});
