--- apps/spectator/components/common/GameCard/pieces/Wrapper.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const gameContext = style({
-  position: 'relative',
-  justifyContent: 'center',
-  height: '100%',
-});
