--- apps/spectator/components/game/RecordList/RecordList.css.ts
@@ -1,11 +1,0 @@
-import { theme } from '@hcc/styles';
-import { styleVariants } from '@vanilla-extract/css';
-
-import { errorFallback as videoErrorFallback } from '../Video/Video.css';
-
-export const recordList = styleVariants({
-  title: { marginBottom: '0.75rem', color: theme.colors.primary[3] },
-  list: { marginInlineStart: '1.25rem', borderInlineStartWidth: 1 },
-});
-
-export const errorFallback = videoErrorFallback;
